rename wavemeter hardware module and remove obsolete logic, gui and interface files
