+++ src/qudi/hardware/wavemeter/high_finesse_wavemeter.py
@@ -1,9 +1,12 @@
 # -*- coding: utf-8 -*-
 
 """
-This module contains a POI Manager core class which gives capability to mark
-points of interest, re-optimise their position, and keep track of sample drift
-over time.
+This file contains the qudi hardware module for the HighFinesse wavemeter. It implements the
+DataInStreamInterface. Communication with the hardware is done via callback functions such that no new data is missed.
+Measurement timestamps provided by the hardware are used for sample timing. As an approximation, the timestamps
+corresponding to the first active channel are equally used for all channels. Considering the fixed round-trip time
+composing the individual channel exposure times and switching times, this implementation should satisfy all
+synchronization needs to an extent that is possible to implement on this software level.
 
 Copyright (c) 2021, the qudi developers. See the AUTHORS.md file at the top-level directory of this
 distribution and on <https://github.com/Ulm-IQO/qudi-iqo-modules/>
@@ -22,272 +25,508 @@ You should have received a copy of the GNU Lesser General Public License along w
 If not, see <https://www.gnu.org/licenses/>.
 """
 
-from qtpy import QtCore
-import ctypes   # is a foreign function library for Python. It provides C
-                # compatible data types, and allows calling functions in DLLs
-                # or shared libraries. It can be used to wrap these libraries
-                # in pure Python.
+import time
+from ctypes import cast, c_double, c_int, c_long, POINTER, windll, WINFUNCTYPE
+from typing import Union, Optional, List, Tuple, Sequence
 
-from qudi.interface.wavemeter_interface import WavemeterInterface
-from qudi.core.module import Base
+import numpy as np
+from PySide2 import QtCore
+
 from qudi.core.configoption import ConfigOption
+import qudi.hardware.wavemeter.high_finesse_constants as high_finesse_constants
 from qudi.util.mutex import Mutex
+from qudi.util.constraints import ScalarConstraint
+from qudi.interface.data_instream_interface import DataInStreamInterface, DataInStreamConstraints, StreamingMode, \
+    SampleTiming
+
+_CALLBACK = WINFUNCTYPE(c_int, c_long, c_long, c_long, c_double, c_long)
 
 
-class HardwarePull(QtCore.QObject):
-    """ Helper class for running the hardware communication in a separate thread. """
-
-    # signal to deliver the wavelength to the parent class
-    sig_wavelength = QtCore.Signal(float, float)
-
-    def __init__(self, parentclass):
-        super().__init__()
-
-        # remember the reference to the parent class to access functions ad settings
-        self._parentclass = parentclass
-
-
-    def handle_timer(self, state_change):
-        """ Threaded method that can be called by a signal from outside to start the timer.
-
-        @param bool state: (True) starts timer, (False) stops it.
-        """
-
-        if state_change:
-            self.timer = QtCore.QTimer()
-            self.timer.timeout.connect(self._measure_thread)
-            self.timer.start(self._parentclass._measurement_timing)
-        else:
-            if hasattr(self, 'timer'):
-                self.timer.stop()
-
-    def _measure_thread(self):
-        """ The threaded method querying the data from the wavemeter.
-        """
-
-        # update as long as the state is busy
-        if self._parentclass.module_state() == 'running':
-            # get the current wavelength from the wavemeter
-            temp1=float(self._parentclass._wavemeterdll.GetWavelength(0))
-            temp2=float(self._parentclass._wavemeterdll.GetWavelength(0))
-
-            # send the data to the parent via a signal
-            self.sig_wavelength.emit(temp1, temp2)
-
-
-
-class HighFinesseWavemeter(WavemeterInterface):
-    """ Hardware class to controls a High Finesse Wavemeter.
+class HighFinesseWavemeter(DataInStreamInterface):
+    """
+    HighFinesse wavelength meter as an in-streaming device.
 
     Example config for copy-paste:
 
-    high_finesse_wavemeter:
-        module.Class: 'high_finesse_wavemeter.HighFinesseWavemeter'
-        fitlogic
-            measurement_timing: 10.0 # in seconds
-
+    wavemeter:
+        module.Class: 'wavemeter.high_finesse_wavemeter.HighFinesseWavemeter'
+        options:
+            channels:
+                red_laser_1:
+                    switch_ch: 1    # channel on the wavemeter switch
+                    unit: 'nm'    # wavelength (nm) or frequency (THz)
+                    medium: 'vac' # for wavelength: air or vac
+                    exposure: 10  # exposure time in ms
+                red_laser_2:
+                    switch_ch: 2
+                    unit: 'nm'
+                    medium: 'vac'
+                    exposure: 10
+                green_laser:
+                    switch_ch: 3
+                    unit: 'nm'
+                    medium: 'vac'
+                    exposure: 10
+                yellow_laser:
+                    switch_ch: 4
+                    unit: 'nm'
+                    medium: 'vac'
+                    exposure: 10
     """
 
+    # declare signals
+    sigNewWavelength = QtCore.Signal(object)
+
     # config options
-    _measurement_timing = ConfigOption('measurement_timing', default=10.)
-
-    # signals
-    sig_handle_timer = QtCore.Signal(bool)
-
-    #############################################
-    # Flags for the external DLL
-    #############################################
-
-    # define constants as flags for the wavemeter
-    _cCtrlStop                   = ctypes.c_uint16(0x00)
-    # this following flag is modified to override every existing file
-    _cCtrlStartMeasurment        = ctypes.c_uint16(0x1002)
-    _cReturnWavelangthAir        = ctypes.c_long(0x0001)
-    _cReturnWavelangthVac        = ctypes.c_long(0x0000)
-
+    _wavemeter_ch_config = ConfigOption(
+        name='channels',
+        default={
+            'default_channel': {'switch_ch': 1, 'unit': 'nm', 'medium': 'vac', 'exposure': 10}
+        },
+        missing='info'
+    )
 
     def __init__(self, *args, **kwargs):
-        self.log.warning("This module has not been tested on the new qudi core."
-                         "Use with caution and contribute bug fixed back, please.")
         super().__init__(*args, **kwargs)
+        self._lock = Mutex()
+        self._callback_function = None
+        self._wavemeterdll = None
 
-        #locking for thread safety
-        self.threadlock = Mutex()
+        # Internal settings
+        self._channel_names = None  # dictionary with switch channel numbers as keys, channel names as values
+        self._channel_buffer_size = 1024**2
+        self._active_switch_channels = None  # list of active switch channel numbers
+        self._unit_return_type = {}
 
-        # the current wavelength read by the wavemeter in nm (vac)
-        self._current_wavelength = 0.0
-        self._current_wavelength2 = 0.0
+        # Data buffer
+        self._wm_start_time = None
+        self._data_buffer = None
+        self._timestamp_buffer = None
+        self._current_buffer_positions = None
+        self._buffer_overflow = False
 
+        # Stored hardware constraints
+        self._constraints = None
 
-    def on_activate(self):
-        #############################################
-        # Initialisation to access external DLL
-        #############################################
+    def on_activate(self) -> None:
         try:
-            # imports the spectrometer specific function from dll
-            self._wavemeterdll = ctypes.windll.LoadLibrary('wlmData.dll')
+            # load wavemeter DLL
+            self._wavemeterdll = windll.LoadLibrary('wlmData.dll')
+        except FileNotFoundError:
+            self.log.error('There is no wavemeter installed on this computer.\n'
+                           'Please install a High Finesse wavemeter and try again.')
+            return
 
-        except:
-            self.log.critical('There is no Wavemeter installed on this '
-                    'Computer.\nPlease install a High Finesse Wavemeter and '
-                    'try again.')
+        # define function header for a later call
+        self._wavemeterdll.Instantiate.argtypes = [c_long, c_long, POINTER(c_long), c_long]
+        self._wavemeterdll.Instantiate.restype = POINTER(c_long)
+        self._wavemeterdll.ConvertUnit.restype = c_double
+        self._wavemeterdll.ConvertUnit.argtypes = [c_double, c_long, c_long]
+        self._wavemeterdll.SetExposureNum.restype = c_long
+        self._wavemeterdll.SetExposureNum.argtypes = [c_long, c_long, c_long]
+        self._wavemeterdll.GetExposureNum.restype = c_long
+        self._wavemeterdll.GetExposureNum.argtypes = [c_long, c_long, c_long]
 
-        # define the use of the GetWavelength function of the wavemeter
-#        self._GetWavelength2 = self._wavemeterdll.GetWavelength2
-        # return data type of the GetWavelength function of the wavemeter
-        self._wavemeterdll.GetWavelength2.restype = ctypes.c_double
-        # parameter data type of the GetWavelength function of the wavemeter
-        self._wavemeterdll.GetWavelength2.argtypes = [ctypes.c_double]
+        # configure wavemeter channels
+        self._channel_names = {}
+        channel_units = {}
+        for ch_name, info in self._wavemeter_ch_config.items():
+            ch = info['switch_ch']
+            unit = info['unit']
+            medium = info['medium']
+            self._channel_names[ch] = ch_name
 
-        # define the use of the GetWavelength function of the wavemeter
-#        self._GetWavelength = self._wavemeterdll.GetWavelength
-        # return data type of the GetWavelength function of the wavemeter
-        self._wavemeterdll.GetWavelength.restype = ctypes.c_double
-        # parameter data type of the GetWavelength function of the wavemeter
-        self._wavemeterdll.GetWavelength.argtypes = [ctypes.c_double]
+            if unit == 'THz' or unit == 'Hz':
+                channel_units[ch_name] = unit
+                self._unit_return_type[ch] = high_finesse_constants.cReturnFrequency
+            elif unit == 'nm' or unit == 'm':
+                channel_units[ch_name] = unit
+                if medium == 'vac':
+                    self._unit_return_type[ch] = high_finesse_constants.cReturnWavelengthVac
+                elif medium == 'air':
+                    self._unit_return_type[ch] = high_finesse_constants.cReturnWavelengthAir
+                else:
+                    self.log.error(f'Invalid medium: {medium}. Valid media are vac and air.')
+                    self._unit_return_type[ch] = high_finesse_constants.cReturnWavelengthVac
+            else:
+                self.log.error(f'Invalid unit: {unit}. Valid units are THz and nm.')
+                channel_units[ch_name] = None
 
-        # define the use of the ConvertUnit function of the wavemeter
-#        self._ConvertUnit = self._wavemeterdll.ConvertUnit
-        # return data type of the ConvertUnit function of the wavemeter
-        self._wavemeterdll.ConvertUnit.restype = ctypes.c_double
-        # parameter data type of the ConvertUnit function of the wavemeter
-        self._wavemeterdll.ConvertUnit.argtypes = [ctypes.c_double, ctypes.c_long, ctypes.c_long]
+            exp_time = info.get('exposure')
+            if exp_time is not None:
+                res = self._wavemeterdll.SetExposureNum(ch, 1, exp_time)
+                if res != 0:
+                    self.log.warning('Wavemeter error while setting exposure time.')
 
-        # manipulate perdefined operations with simple flags
-#        self._Operation = self._wavemeterdll.Operation
-        # return data type of the Operation function of the wavemeter
-        self._wavemeterdll.Operation.restype = ctypes.c_long
-        # parameter data type of the Operation function of the wavemeter
-        self._wavemeterdll.Operation.argtypes = [ctypes.c_ushort]
+        self._active_switch_channels = list(self._channel_names)
 
-        # create an indepentent thread for the hardware communication
-        self.hardware_thread = QtCore.QThread()
+        # set up constraints
+        self._constraints = DataInStreamConstraints(
+            channel_units=channel_units,
+            sample_timing=SampleTiming.TIMESTAMP,
+            # TODO: implement fixed streaming mode
+            streaming_modes=[StreamingMode.CONTINUOUS],
+            data_type=np.float64,
+            channel_buffer_size=ScalarConstraint(default=1024**2,  # 8 MB
+                                                 bounds=(128, 1024**3),  # max = 8 GB
+                                                 increment=1,
+                                                 enforce_int=True),
+        )
 
-        # create an object for the hardware communication and let it live on the new thread
-        self._hardware_pull = HardwarePull(self)
-        self._hardware_pull.moveToThread(self.hardware_thread)
+    def on_deactivate(self) -> None:
+        self.stop_stream()
 
-        # connect the signals in and out of the threaded object
-        self.sig_handle_timer.connect(self._hardware_pull.handle_timer)
-        self._hardware_pull.sig_wavelength.connect(self.handle_wavelength)
+        # free memory
+        self._data_buffer = None
+        self._timestamp_buffer = None
 
-        # start the event loop for the hardware
-        self.hardware_thread.start()
+        # clean up by removing reference to the ctypes library object
+        del self._wavemeterdll
 
+    @property
+    def constraints(self) -> DataInStreamConstraints:
+        """ Read-only property returning the constraints on the settings for this data streamer. """
+        return self._constraints
 
-    def on_deactivate(self):
-        if self.module_state() != 'idle' and self.module_state() != 'deactivated':
-            self.stop_acquisition()
-        self.hardware_thread.quit()
-        self.sig_handle_timer.disconnect()
-        self._hardware_pull.sig_wavelength.disconnect()
+    def _get_callback_ex(self):
+        """
+        Define the callback procedure that should be called by the dll every time a new measurement result
+        is available or any of the wavelength meter's states changes.
+        :return: callback function
+        """
 
-        try:
-            # clean up by removing reference to the ctypes library object
-            del self._wavemeterdll
+        def handle_callback(version, mode, intval, dblval, res1):
+            """
+            Function called upon wavelength meter state change or if a new measurement result is available.
+            See wavemeter manual section on CallbackProc for details.
+
+            In this implementation, the new wavelength is converted to the desired unit and
+            appended to a list together with the current timestamp.
+
+            :param version: Device version number which called the procedure.
+            Only relevant if multiple wavemeter applications are running.
+            :param mode: Indicates which state has changed or what new result is available.
+            :param intval: Contains the time stamp rounded to ms if mode indicates that the new value is in dblval.
+            If not, it contains the new value itself.
+            :param dblval: May contain the new value (e.g. wavelength), depending on mode.
+            :param res1: Mostly meaningless.
+            :return: 0
+            """
+            if mode == high_finesse_constants.cmiOperation and intval == high_finesse_constants.cStop:
+                self.log.error('Wavemeter acquisition was stopped during stream.')
+                return 0
+
+            if self._buffer_overflow:
+                return 0
+
+            # TODO: why does this not work in the lock anymore?
+            # with self._lock:
+            if True:
+                # see if new data is from one of the active channels
+                ch = high_finesse_constants.cmi_wavelength_n.get(mode)
+                if ch in self._active_switch_channels:
+                    i = self._active_switch_channels.index(ch)
+
+                    if self._current_buffer_positions[i] >= self.channel_buffer_size:
+                        self._buffer_overflow = True
+                        raise OverflowError(
+                            'Streaming buffer encountered an overflow while receiving a callback from the wavemeter. '
+                            'Please increase the buffer size or speed up data reading.'
+                        )
+
+                    # wavemeter records timestamps in ms
+                    timestamp = 1e-3 * intval
+                    # in case this is the first recorded measurement,
+                    # set the timing offset to the start of the stream
+                    if self._wm_start_time is None:
+                        self._wm_start_time = timestamp
+                    timestamp -= self._wm_start_time
+                    # unit conversion
+                    converted_value = self._wavemeterdll.ConvertUnit(
+                        dblval, high_finesse_constants.cReturnWavelengthVac, self._unit_return_type[ch]
+                    )
+
+                    # insert the new data into the buffers
+                    self._data_buffer[i, self._current_buffer_positions[i]] = converted_value
+                    if i == 0:
+                        # only record the timestamp of the first active channel
+                        self._timestamp_buffer[self._current_buffer_positions[0]] = timestamp
+                    self._current_buffer_positions[i] += 1
+
+                    self.sigNewWavelength.emit(converted_value)
             return 0
-        except:
-            self.log.error('Could not unload the wlmData.dll of the '
-                    'wavemeter.')
 
+        self._callback_function = _CALLBACK(handle_callback)
+        return self._callback_function
 
-    #############################################
-    # Methods of the main class
-    #############################################
+    def start_stream(self) -> None:
+        """ Start the data acquisition/streaming """
+        with self._lock:
+            if self.module_state() == 'idle':
+                self.module_state.lock()
 
-    def handle_wavelength(self, wavelength1, wavelength2):
-        """ Function to save the wavelength, when it comes in with a signal.
+                # start callback procedure
+                self._wavemeterdll.Instantiate(
+                    high_finesse_constants.cInstNotification,  # long ReasonForCall
+                    high_finesse_constants.cNotifyInstallCallbackEx,  # long Mode
+                    cast(self._get_callback_ex(), POINTER(c_long)),  # long P1: function
+                    0)  # long P2: callback thread priority, 0 = standard
+
+                self._init_buffers()
+            else:
+                self.log.warning('Unable to start input stream. It is already running.')
+
+    def stop_stream(self) -> None:
+        """ Stop the data acquisition/streaming """
+        with self._lock:
+            if self.module_state() == 'locked':
+                self._wavemeterdll.Instantiate(
+                    high_finesse_constants.cInstNotification,  # long ReasonForCall
+                    high_finesse_constants.cNotifyRemoveCallback,  # long mode
+                    cast(self._callback_function, POINTER(c_long)),
+                    # long P1: function
+                    0)  # long P2: callback thread priority, 0 = standard
+                self._callback_function = None
+                self._wm_start_time = None
+
+                self.module_state.unlock()
+            else:
+                self.log.warning('Unable to stop wavemeter input stream as nothing is running.')
+
+    def read_data_into_buffer(self,
+                              data_buffer: np.ndarray,
+                              timestamp_buffer: Optional[np.ndarray] = None) -> None:
+        """ Read data from the stream buffer into a 1D numpy array given as parameter.
+        All samples for each channel are stored in consecutive blocks one after the other.
+        The data_buffer can be unraveled into channel samples with:
+
+            data_buffer.reshape([<channel_count>, <samples_per_channel>])
+
+        The data_buffer array must have the same data type as self.constraints.data_type.
+
+        In case of SampleTiming.TIMESTAMP a 1D numpy.float64 timestamp_buffer array has to be
+        provided to be filled with timestamps corresponding to the data_buffer array. It must be
+        exactly of length:
+
+            data_buffer // <channel_count>
+
+        This function is blocking until the entire buffer has been filled.
         """
-        self._current_wavelength = wavelength1
-        self._current_wavelength2 = wavelength2
+        with self._lock:
+            if self.module_state() != 'locked':
+                raise RuntimeError('Unable to read data. Stream is not running.')
 
-    def start_acquisition(self):
-        """ Method to start the wavemeter software.
+            number_of_channels, number_of_samples = self._validate_buffers(data_buffer, timestamp_buffer)
+            total_samples = number_of_samples * number_of_channels
 
-        @return int: error code (0:OK, -1:error)
+            # wait until requested number of samples is available
+            while self.available_samples < number_of_samples:
+                if self.module_state() != 'locked':
+                    break
+                # wait for 10 ms
+                time.sleep(0.01)
 
-        Also the actual threaded method for getting the current wavemeter reading is started.
+            data_buffer[:total_samples] = self._data_buffer[:, :number_of_samples].flatten()
+            timestamp_buffer[:total_samples] = self._timestamp_buffer[:number_of_samples]
+            self._remove_samples_from_buffer(number_of_samples)
+
+    def read_available_data_into_buffer(self,
+                                        data_buffer: np.ndarray,
+                                        timestamp_buffer: Optional[np.ndarray] = None) -> int:
+        """ Read data from the stream buffer into a 1D numpy array given as parameter.
+        All samples for each channel are stored in consecutive blocks one after the other.
+        The number of samples read per channel is returned and can be used to slice out valid data
+        from the buffer arrays like:
+
+            valid_data = data_buffer[:<channel_count> * <return_value>]
+            valid_timestamps = timestamp_buffer[:<return_value>]
+
+        See "read_data_into_buffer" documentation for more details.
+
+        This method will read all currently available samples into buffer. If number of available
+        samples exceeds buffer size, read only as many samples as fit into the buffer.
         """
+        with self._lock:
+            if self.module_state() != 'locked':
+                raise RuntimeError('Unable to read data. Stream is not running.')
 
-        # first check its status
-        if self.module_state() == 'running':
-            self.log.error('Wavemeter busy')
-            return -1
+            number_of_channels, requested_samples = self._validate_buffers(data_buffer, timestamp_buffer)
+            read_samples = number_of_channels * min(requested_samples, self.available_samples)
 
+            data_buffer[:read_samples] = self._data_buffer[:, :read_samples].flatten()
+            timestamp_buffer[:read_samples] = self._timestamp_buffer[:read_samples]
+            self._remove_samples_from_buffer(read_samples)
 
-        self.module_state.run()
-        # actually start the wavemeter
-        self._wavemeterdll.Operation(self._cCtrlStartMeasurment) #starts measurement
+        return read_samples
 
-        # start the measuring thread
-        self.sig_handle_timer.emit(True)
+    def read_data(self,
+                  number_of_samples: Optional[int] = None
+                  ) -> Tuple[np.ndarray, Union[np.ndarray, None]]:
+        """ Read data from the stream buffer into a 1D numpy array and return it.
+        All samples for each channel are stored in consecutive blocks one after the other.
+        The returned data_buffer can be unraveled into channel samples with:
 
-        return 0
+            data_buffer.reshape([<channel_count>, number_of_samples])
 
-    def stop_acquisition(self):
-        """ Stops the Wavemeter from measuring and kills the thread that queries the data.
+        The numpy array data type is the one defined in self.constraints.data_type.
 
-        @return int: error code (0:OK, -1:error)
+        In case of SampleTiming.TIMESTAMP a 1D numpy.float64 timestamp_buffer array will be
+        returned as well with timestamps corresponding to the data_buffer array.
+
+        If number_of_samples is omitted all currently available samples are read from buffer.
+        This method will not return until all requested samples have been read or a timeout occurs.
         """
-        # check status just for a sanity check
-        if self.module_state() == 'idle':
-            self.log.warning('Wavemeter was already stopped, stopping it '
-                    'anyway!')
+        with self._lock:
+            if self.module_state() != 'locked':
+                self.log.error('Unable to read data. Device is not running.')
+
+            read_samples = number_of_samples if number_of_samples is not None else self.available_samples
+
+            data_buffer = np.empty(read_samples * len(self.active_channels), dtype=self.constraints.data_type)
+            timestamp_buffer = np.emtpy(read_samples, dtype=np.float64)
+            self.read_data_into_buffer(data_buffer, timestamp_buffer)
+
+        return data_buffer, timestamp_buffer
+
+    def read_single_point(self) -> Tuple[np.ndarray, Union[None, np.float64]]:
+        """ This method will initiate a single sample read on each configured data channel.
+        In general this sample may not be acquired simultaneous for all channels and timing in
+        general can not be assured. Us this method if you want to have a non-timing-critical
+        snapshot of your current data channel input.
+        May not be available for all devices.
+        The returned 1D numpy array will contain one sample for each channel.
+
+        In case of SampleTiming.TIMESTAMP a single numpy.float64 timestamp value will be returned
+        as well.
+        """
+        if self.module_state() != 'locked':
+            self.log.error('Unable to read data. Device is not running.')
+            return np.empty(0, dtype=self.constraints.data_type), None
+
+        i = self._current_buffer_positions.min() - 1
+        data = self._data_buffer[:, i]
+        timestamp = self._timestamp_buffer[i]
+        return data, timestamp
+
+    @property
+    def sample_rate(self) -> float:
+        """ Read-only property returning the currently set sample rate in Hz.
+        For SampleTiming.CONSTANT this is the sample rate of the hardware, for any other timing mode
+        this property represents only a hint to the actual hardware timebase and can not be
+        considered accurate.
+
+        For the wavemeter, it is estimated by the exposure times per channel and switching times if
+        more than one channel is active.
+        """
+        exposure_times = []
+        for ch in self._active_switch_channels:
+            t = self._wavemeterdll.GetExposureNum(ch, 1, 0)
+            exposure_times.append(t)
+        total_exposure_time = sum(exposure_times)
+
+        switching_time = 12
+        n_channels = len(self._active_switch_channels)
+        if n_channels > 1:
+            turnaround_time_ms = total_exposure_time + n_channels * switching_time
         else:
-            # stop the measurement thread
-            self.sig_handle_timer.emit(True)
-            # set status to idle again
-            self.module_state.stop()
+            turnaround_time_ms = total_exposure_time
 
-        # Stop the actual wavemeter measurement
-        self._wavemeterdll.Operation(self._cCtrlStop)
+        sample_rate = 1e3 / turnaround_time_ms
+        return sample_rate
 
-        return 0
+    @property
+    def streaming_mode(self) -> StreamingMode:
+        """ Read-only property returning the currently configured StreamingMode Enum """
+        return StreamingMode.CONTINUOUS
 
-    def get_current_wavelength(self, kind="air"):
-        """ This method returns the current wavelength.
+    @property
+    def active_channels(self) -> List[str]:
+        """ Read-only property returning the currently configured active channel names """
+        ch_names = [self._channel_names[ch] for ch in self._active_switch_channels]
+        return ch_names
 
-        @param string kind: can either be "air" or "vac" for the wavelength in air or vacuum, respectively.
-
-        @return float: wavelength (or negative value for errors)
+    @property
+    def available_samples(self) -> int:
         """
-        if kind in "air":
-            # for air we need the convert the current wavelength. The Wavemeter DLL already gives us a nice tool do do so.
-            return float(self._wavemeterdll.ConvertUnit(self._current_wavelength,self._cReturnWavelangthVac,self._cReturnWavelangthAir))
-        if kind in "vac":
-            # for vacuum just return the current wavelength
-            return float(self._current_wavelength)
-        return -2.0
-
-    def get_current_wavelength2(self, kind="air"):
-        """ This method returns the current wavelength of the second input channel.
-
-        @param string kind: can either be "air" or "vac" for the wavelength in air or vacuum, respectively.
-
-        @return float: wavelength (or negative value for errors)
+        Read-only property to return the currently available number of samples per channel ready
+        to read from buffer.
         """
-        if kind in "air":
-            # for air we need the convert the current wavelength. The Wavemeter DLL already gives us a nice tool do do so.
-            return float(self._wavemeterdll.ConvertUnit(self._current_wavelength2,self._cReturnWavelangthVac,self._cReturnWavelangthAir))
-        if kind in "vac":
-            # for vacuum just return the current wavelength
-            return float(self._current_wavelength2)
-        return -2.0
+        if self.module_state() != 'locked':
+            return 0
 
-    def get_timing(self):
-        """ Get the timing of the internal measurement thread.
+        # all channels must have been read out in order to count as an available sample
+        return min(self._current_buffer_positions)
 
-        @return float: clock length in second
+    @property
+    def channel_buffer_size(self) -> int:
+        """ Read-only property returning the currently set buffer size in samples per channel.
+        The total buffer size in bytes can be estimated by:
+            <buffer_size> * <channel_count> * numpy.nbytes[<data_type>]
+
+        For StreamingMode.FINITE this will also be the total number of samples to acquire per
+        channel.
         """
-        return self._measurement_timing
+        return self._channel_buffer_size
 
-    def set_timing(self, timing):
-        """ Set the timing of the internal measurement thread.
+    def configure(self,
+                  active_channels: Sequence[str],
+                  streaming_mode: Union[StreamingMode, int],
+                  channel_buffer_size: int,
+                  sample_rate: float) -> None:
+        """ Configure a data stream. See read-only properties for information on each parameter. """
+        if self.module_state() == 'locked':
+            raise RuntimeError('Unable to configure data stream while it is already running')
 
-        @param float timing: clock length in second
+        if active_channels is not None:
+            self._active_switch_channels = []
+            for ch in active_channels:
+                if ch in self._wavemeter_ch_config:
+                    self._active_switch_channels.append(self._wavemeter_ch_config[ch]['switch_ch'])
+                else:
+                    self.log.error(f'Channel {ch} is not set up in the config file. Available channels '
+                                   f'are {list(self._channel_names.keys())}.')
 
-        @return int: error code (0:OK, -1:error)
+        if streaming_mode is not None and streaming_mode != StreamingMode.CONTINUOUS:
+            self.log.warning('Only continuous streaming is supported, ignoring this setting.')
+
+        if channel_buffer_size is not None:
+            self.constraints.channel_buffer_size.is_valid(channel_buffer_size)
+            self._channel_buffer_size = channel_buffer_size
+
+    def _init_buffers(self) -> None:
+        """ Initialize buffers and the current buffer position marker. """
+        n = len(self._active_switch_channels)
+        self._data_buffer = np.zeros([n, self._channel_buffer_size], dtype=self.constraints.data_type)
+        self._timestamp_buffer = np.zeros(self._channel_buffer_size, dtype=np.float64)
+        self._current_buffer_positions = np.zeros(n, dtype=int)
+        self._buffer_overflow = False
+
+    def _remove_samples_from_buffer(self, number_of_samples: int) -> None:
         """
-        self._measurement_timing=float(timing)
-        return 0
+        Remove samples that have been read from buffer to make space for new samples.
+        :param number_of_samples: number of samples to clear off the buffer
+        :return: None
+        """
+        self._data_buffer = np.roll(self._data_buffer, -number_of_samples, axis=1)
+        self._timestamp_buffer = np.roll(self._timestamp_buffer, -number_of_samples)
+        self._current_buffer_positions -= number_of_samples
 
+    def _validate_buffers(self,
+                          data_buffer: np.ndarray,
+                          timestamp_buffer: np.ndarray) -> int:
+        """ Validate arguments for read_[available]_data_into_buffer methods. """
+        if not isinstance(data_buffer, np.ndarray) or data_buffer.dtype != self.constraints.data_type:
+            self.log.error(f'data_buffer must be numpy.ndarray with dtype {self.constraints.data_type}.')
+
+        if not isinstance(timestamp_buffer, np.ndarray) or timestamp_buffer.dtype != np.float64:
+            print(timestamp_buffer.dtype)
+            self.log.error(f'timestamp_buffer must be provided for the wavemeter and '
+                           f'it must be a numpy.ndarray with dtype np.float64.')
+
+        number_of_channels = len(self.active_channels)
+        number_of_samples = data_buffer.size // number_of_channels
+
+        if timestamp_buffer.size != number_of_samples:
+            self.log.error(f'timestamp_buffer must be exactly of length data_buffer // <channel_count>')
+
+        return number_of_channels, number_of_samples
