--- src/qudi/wavemter_rework.cfg
@@ -1,119 +1,0 @@
-#  Qudi configuration file for the new core.
-#
-
-global:
-    # list of modules to load when starting
-    startup: []
-
-    remote_modules_server:
-        address: 'localhost'
-        port: 12345
-
-    namespace_server_port: 18861
-
-    default_data_dir: 'C:\\Data'
-
-    ## For controlling the appearance of the GUI:
-    stylesheet: 'qdark.qss'
-
-gui:
-    time_series_gui:
-        module.Class: 'time_series.time_series_gui.TimeSeriesGui'
-        #module.Class: 'time_series.time_series_gui_histogram.TimeSeriesGui'
-        options:
-            use_antialias: True  # optional, set to False if you encounter performance issues
-        connect:
-            _time_series_logic_con: time_series_reader_logic
-
-    wavemeter_histogram_gui:
-        module.Class: 'wavemeter.wavemeter_histogram_gui_2.WavemeterHistogramGui'
-        connect:
-            _wavemeter_histogram_logic_con: wavemeter_histogram_logic
-
-    spectrometer:
-        module.Class: 'spectrometer.spectrometer_gui.SpectrometerGui'
-        connect:
-            spectrometer_logic: 'spectrometerlogic'
-
-    qdplotter:
-        module.Class: 'qdplot.qdplot_gui.QDPlotterGui'
-        options:
-            pen_color_list: [[235,130,39], #orange
-                         [97,161,74],  #green
-                         [217,94,200], #pink
-                         [85,151,206], #blue
-                         [216,34,41], #red
-                         [210, 203, 15], #yellow
-                         [153,153,153] #grey
-                         ]
-        connect:
-            qdplot_logic: 'qdplotlogic'
-
-logic:
-    time_series_reader_logic:
-        module.Class: 'time_series_reader_logic.TimeSeriesReaderLogic'
-        options:
-            max_frame_rate: 20  # optional (10Hz by default)
-            calc_digital_freq: True  # optional (True by default)
-        connect:
-            streamer: instream_dummy # also connect to instream_dummy ?both?
-
-    wavemeter_histogram_logic:
-        module.Class: 'wavemeter_histogram_logic.WavemeterLogic'
-        options:
-            max_frame_rate: 20  # optional (10Hz by default)
-            calc_digital_freq: False  # optional (True by default)
-        connect:
-            streamer: wave_instreamer # also connect to instream_dummy ?both?
-            counterlogic: time_series_reader_logic
-
-    spectrometerlogic:
-        module.Class: 'spectrometer_logic.SpectrometerLogic'
-        connect:
-            spectrometer: 'myspectrometer'
-
-    qdplotlogic:
-        module.Class: 'qdplot_logic.QDPlotLogic'
-        options:
-            default_plot_number: 3
-
-hardware:
-    mywavemeter:
-        module.Class: 'dummy.wavemeter_dummy.WavemeterDummy'
-        options:
-            measurement_timing: 10.0e-3 # measurement timing should be given in seconds
-            unit: 'vac' # unit of the read out wavelength
-            automatic_acquisition: True # if set True on activation of the module measurement starts
-
-    instream_dummy:
-        module.Class: 'dummy.data_instream_dummy.InStreamDummy'
-        options:
-            digital_channels:  # optional, must provide at least one digital or analog channel
-                - 'digital 1'
-                #- 'digital 2'
-                #- 'digital 3'
-            #analog_channels:  # optional, must provide at least one digital or analog channel
-                #- 'analog 1'
-                #- 'analog 2'
-            digital_event_rates:  # optional, must have as many entries as digital_channels or just one
-                - 1000
-                #- 10000
-                #- 100000
-            # digital_event_rates: 100000
-            #analog_amplitudes:  # optional, must have as many entries as analog_channels or just one
-                #- 5
-                #- 10
-            # analog_amplitudes: 10  # optional (10V by default)
-
-    wave_instreamer:
-        module.Class: 'wavemeter.wavemeter_as_instreamer.WavemeterAsInstreamer'
-        options:
-            channels:
-                1:
-                    unit: 'nm'    # wavelength (nm) or frequency (THz)
-                    medium: 'vac' # for wavelength: air or vac
-                #     exposure: 10  # exposure time in ms
-                # 2:
-                #     unit: 'nm'
-                #     medium: 'vac'
-                #     exposure: 10
